Change task name from 2D registration to registration
