+++ tests/tasks/test_registration.py
@@ -28,8 +28,8 @@ from fractal_tasks_core.tasks.apply_registration_to_image import (
 from fractal_tasks_core.tasks.apply_registration_to_ROI_tables import (
     apply_registration_to_ROI_tables,
 )
-from fractal_tasks_core.tasks.calculate_registration_image_based import (
-    calculate_registration_image_based,
+from fractal_tasks_core.tasks.calculate_2D_registration_image_based import (
+    calculate_2D_registration_image_based,
 )
 from fractal_tasks_core.tasks.cellpose_segmentation import (
     cellpose_segmentation,
@@ -77,6 +77,13 @@ def _shift_image(
     new_array[:, :-shift_x_pxl] = new_array[:, shift_x_pxl:]
     new_array[:-shift_y_pxl] = new_array[shift_y_pxl:, :]
 
+    # Replace offset stripes with random values
+    # new_array[:, -shift_x_pxl:] = np.random.randint(
+    #     0, 1000, size=new_array[:, -shift_x_pxl:].shape
+    # )
+    # new_array[-shift_y_pxl:, :] = np.random.randint(
+    #     0, 1000, size=new_array[-shift_y_pxl:, :].shape
+    # )
     new_array[:, -shift_x_pxl:] = (
         np.ones(new_array[:, -shift_x_pxl:].shape) * 110
     )
@@ -263,7 +270,7 @@ def test_multiplexing_registration(
 
     # Calculate registration
     for component in metadata["image"]:
-        calculate_registration_image_based(
+        calculate_2D_registration_image_based(
             input_paths=[str(zarr_path_mip)],
             output_path=str(zarr_path_mip),
             metadata=metadata,
